+++ src/View/Components/EdgeToolbar.php
@@ -2,10 +2,14 @@
 
 namespace ArtisanFlow\WireFlow\View\Components;
 
+use ArtisanFlow\WireFlow\Concerns\ValidatesEnumProps;
+use ArtisanFlow\WireFlow\Enums\ToolbarShow;
 use Illuminate\View\Component;
 
 class EdgeToolbar extends Component
 {
+    use ValidatesEnumProps;
+
     public string $directive;
 
     public function __construct(
@@ -13,6 +17,8 @@ class EdgeToolbar extends Component
         public bool $below = false,
         public string $show = 'selected',
     ) {
+        self::validateEnum(ToolbarShow::class, $show, 'show');
+
         $this->directive = 'x-flow-edge-toolbar';
         if ($this->below) {
             $this->directive .= '.below';